+++ src/main/java/com/box/sdk/BoxFile.java
@@ -1,7 +1,11 @@
 package com.box.sdk;
 
+import static com.eclipsesource.json.Json.NULL;
+
 import com.box.sdk.http.HttpMethod;
 import com.box.sdk.internal.utils.Parsers;
+import com.box.sdk.sharedlink.BoxSharedLinkRequest;
+import com.eclipsesource.json.Json;
 import com.eclipsesource.json.JsonArray;
 import com.eclipsesource.json.JsonObject;
 import com.eclipsesource.json.JsonValue;
@@ -132,16 +136,25 @@ public class BoxFile extends BoxItem {
         return FILE_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID());
     }
 
+    /**
+     * Creates a new shared link for this item.
+     *
+     * <p>This method is a convenience method for manually creating a new shared link and applying it to this item with
+     * {@link BoxItem.Info#setSharedLink}. You may want to create the shared link manually so that it can be updated along with
+     * other changes to the item's info in a single network request, giving a boost to performance.</p>
+     *
+     * @param access      the access level of the shared link.
+     * @param unshareDate the date and time at which the link will expire. Can be null to create a non-expiring link.
+     * @param permissions the permissions of the shared link. Can be null to use the default permissions.
+     * @return the created shared link.
+     * @deprecated use {@link BoxFile#createSharedLink(BoxSharedLinkRequest)}
+     */
     @Override
+    @Deprecated
     public BoxSharedLink createSharedLink(BoxSharedLink.Access access, Date unshareDate,
                                           BoxSharedLink.Permissions permissions) {
 
-        BoxSharedLink sharedLink = new BoxSharedLink(access, unshareDate, permissions);
-        Info info = new Info();
-        info.setSharedLink(sharedLink);
-
-        this.updateInfo(info);
-        return info.getSharedLink();
+        return createSharedLink(new BoxSharedLink(access, unshareDate, permissions));
     }
 
     /**
@@ -152,11 +165,26 @@ public class BoxFile extends BoxItem {
      * @param permissions The permissions to set on the shared link for the Box file.
      * @param password    Password set on the shared link to give access to the Box file.
      * @return information about the newly created shared link.
+     * @deprecated Use {@link BoxFile#createSharedLink(BoxSharedLinkRequest)}
      */
+    @Deprecated
     public BoxSharedLink createSharedLink(BoxSharedLink.Access access, Date unshareDate,
                                           BoxSharedLink.Permissions permissions, String password) {
 
-        BoxSharedLink sharedLink = new BoxSharedLink(access, unshareDate, permissions, password);
+        return createSharedLink(new BoxSharedLink(access, unshareDate, permissions, password));
+    }
+
+    /**
+     * Creates a shared link.
+     *
+     * @param sharedLinkRequest Shared link to create
+     * @return Created shared link.
+     */
+    public BoxSharedLink createSharedLink(BoxSharedLinkRequest sharedLinkRequest) {
+        return createSharedLink(sharedLinkRequest.asSharedLink());
+    }
+
+    private BoxSharedLink createSharedLink(BoxSharedLink sharedLink) {
         Info info = new Info();
         info.setSharedLink(sharedLink);
 
@@ -201,7 +229,7 @@ public class BoxFile extends BoxItem {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "POST");
         request.setBody(requestJSON.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         BoxComment addedComment = new BoxComment(this.getAPI(), responseJSON.get("id").asString());
         return addedComment.new Info(responseJSON);
@@ -255,7 +283,7 @@ public class BoxFile extends BoxItem {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "POST");
         request.setBody(requestJSON.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         BoxTask addedTask = new BoxTask(this.getAPI(), responseJSON.get("id").asString());
         return addedTask.new Info(responseJSON);
@@ -347,10 +375,9 @@ public class BoxFile extends BoxItem {
         URL url = CONTENT_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID());
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         if (rangeEnd > 0) {
-            request.addHeader("Range", String.format("bytes=%s-%s", Long.toString(rangeStart),
-                Long.toString(rangeEnd)));
+            request.addHeader("Range", String.format("bytes=%s-%s", rangeStart, rangeEnd));
         } else {
-            request.addHeader("Range", String.format("bytes=%s-", Long.toString(rangeStart)));
+            request.addHeader("Range", String.format("bytes=%s-", rangeStart));
         }
 
         BoxAPIResponse response = request.send();
@@ -391,7 +418,7 @@ public class BoxFile extends BoxItem {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "POST");
         request.setBody(copyInfo.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
         BoxFile copiedFile = new BoxFile(this.getAPI(), responseJSON.get("id").asString());
         return copiedFile.new Info(responseJSON);
     }
@@ -427,7 +454,7 @@ public class BoxFile extends BoxItem {
 
         request.setBody(updateInfo.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
         BoxFile movedFile = new BoxFile(this.getAPI(), responseJSON.get("id").asString());
         return movedFile.new Info(responseJSON);
     }
@@ -478,7 +505,7 @@ public class BoxFile extends BoxItem {
     public BoxFile.Info getInfoWithRepresentations(String representationHints, String... fields) {
         if (representationHints.matches(Representation.X_REP_HINTS_PATTERN)) {
             //Since the user intends to get representations, add it to fields, even if user has missed it
-            Set<String> fieldsSet = new HashSet<String>(Arrays.asList(fields));
+            Set<String> fieldsSet = new HashSet<>(Arrays.asList(fields));
             fieldsSet.add("representations");
             String queryString = new QueryStringBuilder().appendParam("fields",
                 fieldsSet.toArray(new String[fieldsSet.size()])).toString();
@@ -524,10 +551,7 @@ public class BoxFile extends BoxItem {
         String repState = representation.getStatus().getState();
 
         if (repState.equals("viewable") || repState.equals("success")) {
-
-            this.makeRepresentationContentRequest(representation.getContent().getUrlTemplate(),
-                assetPath, output);
-            return;
+            this.makeRepresentationContentRequest(representation.getContent().getUrlTemplate(), assetPath, output);
         } else if (repState.equals("pending") || repState.equals("none")) {
 
             String repContentURLString = null;
@@ -536,13 +560,9 @@ public class BoxFile extends BoxItem {
             }
 
             this.makeRepresentationContentRequest(repContentURLString, assetPath, output);
-            return;
-
         } else if (repState.equals("error")) {
-
             throw new BoxAPIException("Representation had error status");
         } else {
-
             throw new BoxAPIException("Representation had unknown status");
         }
 
@@ -622,7 +642,7 @@ public class BoxFile extends BoxItem {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "PUT");
         request.setBody(info.getPendingChanges());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
         info.update(jsonObject);
     }
 
@@ -637,9 +657,9 @@ public class BoxFile extends BoxItem {
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
 
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
         JsonArray entries = jsonObject.get("entries").asArray();
-        Collection<BoxFileVersion> versions = new ArrayList<BoxFileVersion>();
+        Collection<BoxFileVersion> versions = new ArrayList<>();
         for (JsonValue entry : entries) {
             versions.add(new BoxFileVersion(this.getAPI(), entry.asObject(), this.getID()));
         }
@@ -793,7 +813,6 @@ public class BoxFile extends BoxItem {
     public void uploadVersion(InputStream fileContent, String fileContentSHA1, Date modified, long fileSize,
                               ProgressListener listener) {
         this.uploadNewVersion(fileContent, fileContentSHA1, modified, fileSize, listener);
-        return;
     }
 
     /**
@@ -1003,10 +1022,10 @@ public class BoxFile extends BoxItem {
         URL url = GET_COMMENTS_URL_TEMPLATE.build(this.getAPI().getBaseURL(), this.getID());
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         int totalCount = responseJSON.get("total_count").asInt();
-        List<BoxComment.Info> comments = new ArrayList<BoxComment.Info>(totalCount);
+        List<BoxComment.Info> comments = new ArrayList<>(totalCount);
         JsonArray entries = responseJSON.get("entries").asArray();
         for (JsonValue value : entries) {
             JsonObject commentJSON = value.asObject();
@@ -1027,15 +1046,15 @@ public class BoxFile extends BoxItem {
     public List<BoxTask.Info> getTasks(String... fields) {
         QueryStringBuilder builder = new QueryStringBuilder();
         if (fields.length > 0) {
-            builder.appendParam("fields", fields).toString();
+            builder.appendParam("fields", fields);
         }
         URL url = GET_TASKS_URL_TEMPLATE.buildWithQuery(this.getAPI().getBaseURL(), builder.toString(), this.getID());
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         int totalCount = responseJSON.get("total_count").asInt();
-        List<BoxTask.Info> tasks = new ArrayList<BoxTask.Info>(totalCount);
+        List<BoxTask.Info> tasks = new ArrayList<>(totalCount);
         JsonArray entries = responseJSON.get("entries").asArray();
         for (JsonValue value : entries) {
             JsonObject taskJSON = value.asObject();
@@ -1083,7 +1102,7 @@ public class BoxFile extends BoxItem {
         request.addHeader("Content-Type", "application/json");
         request.setBody(metadata.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -1095,7 +1114,7 @@ public class BoxFile extends BoxItem {
      * @return the metadata returned from the server.
      */
     public Metadata setMetadata(String templateName, String scope, Metadata metadata) {
-        Metadata metadataValue = null;
+        Metadata metadataValue;
 
         try {
             metadataValue = this.createMetadata(templateName, scope, metadata);
@@ -1110,7 +1129,7 @@ public class BoxFile extends BoxItem {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
                             value.asObject().get("value").asString());
                     } else if (value.asObject().get("value").isArray()) {
-                        ArrayList<String> list = new ArrayList<String>();
+                        ArrayList<String> list = new ArrayList<>();
                         for (JsonValue jsonValue : value.asObject().get("value").asArray()) {
                             list.add(jsonValue.asString());
                         }
@@ -1162,7 +1181,7 @@ public class BoxFile extends BoxItem {
      */
     public String setClassification(String classificationType) {
         Metadata metadata = new Metadata().add(Metadata.CLASSIFICATION_KEY, classificationType);
-        Metadata classification = null;
+        Metadata classification;
 
         try {
             classification = this.createMetadata(Metadata.CLASSIFICATION_TEMPLATE_KEY, "enterprise", metadata);
@@ -1190,7 +1209,7 @@ public class BoxFile extends BoxItem {
             metadata = this.getMetadata(Metadata.CLASSIFICATION_TEMPLATE_KEY);
 
         } catch (BoxAPIException e) {
-            JsonObject responseObject = JsonObject.readFrom(e.getResponse());
+            JsonObject responseObject = Json.parse(e.getResponse()).asObject();
             String code = responseObject.get("code").asString();
 
             if (e.getResponseCode() == 404 && code.equals("instance_not_found")) {
@@ -1264,9 +1283,9 @@ public class BoxFile extends BoxItem {
 
         BoxJSONResponse response = (BoxJSONResponse) request.send();
 
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
         JsonValue lockValue = responseJSON.get("lock");
-        JsonObject lockJSON = JsonObject.readFrom(lockValue.toString());
+        JsonObject lockJSON = Json.parse(lockValue.toString()).asObject();
 
         return new BoxLock(lockJSON, this.getAPI());
     }
@@ -1280,7 +1299,7 @@ public class BoxFile extends BoxItem {
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "PUT");
 
         JsonObject lockObject = new JsonObject();
-        lockObject.add("lock", JsonObject.NULL);
+        lockObject.add("lock", NULL);
 
         request.setBody(lockObject.toString());
         request.send();
@@ -1327,7 +1346,7 @@ public class BoxFile extends BoxItem {
         URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), scope, typeName);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -1352,7 +1371,7 @@ public class BoxFile extends BoxItem {
         request.addHeader("Content-Type", "application/json-patch+json");
         request.setBody(metadata.getPatch());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -1431,7 +1450,7 @@ public class BoxFile extends BoxItem {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "PUT");
         request.setBody(infoJSON.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
         return new Info(jsonObject);
     }
 
@@ -1453,7 +1472,7 @@ public class BoxFile extends BoxItem {
         request.setBody(body.toString());
 
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
 
         String sessionId = jsonObject.get("id").asString();
         BoxFileUploadSession session = new BoxFileUploadSession(this.getAPI(), sessionId);
@@ -2005,7 +2024,7 @@ public class BoxFile extends BoxItem {
         }
 
         private List<String> parseAllowedInviteeRoles(JsonArray jsonArray) {
-            List<String> roles = new ArrayList<String>(jsonArray.size());
+            List<String> roles = new ArrayList<>(jsonArray.size());
             for (JsonValue value : jsonArray) {
                 roles.add(value.asString());
             }