+++ src/main/java/com/box/sdk/BoxFolder.java
@@ -1,6 +1,8 @@
 package com.box.sdk;
 
 import com.box.sdk.internal.utils.Parsers;
+import com.box.sdk.sharedlink.BoxSharedLinkRequest;
+import com.eclipsesource.json.Json;
 import com.eclipsesource.json.JsonArray;
 import com.eclipsesource.json.JsonObject;
 import com.eclipsesource.json.JsonValue;
@@ -212,16 +214,25 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         return BoxCollaboration.create(this.getAPI(), accessibleByField, itemField, role, notify, canViewPath);
     }
 
+    /**
+     * Creates a new shared link for this item.
+     *
+     * <p>This method is a convenience method for manually creating a new shared link and applying it to this item with
+     * {@link BoxItem.Info#setSharedLink}. You may want to create the shared link manually so that it can be updated along with
+     * other changes to the item's info in a single network request, giving a boost to performance.</p>
+     *
+     * @param access      the access level of the shared link.
+     * @param unshareDate the date and time at which the link will expire. Can be null to create a non-expiring link.
+     * @param permissions the permissions of the shared link. Can be null to use the default permissions.
+     * @return the created shared link.
+     * @deprecated use {@link BoxFolder#createSharedLink(BoxSharedLinkRequest)}
+     */
     @Override
+    @Deprecated
     public BoxSharedLink createSharedLink(BoxSharedLink.Access access, Date unshareDate,
                                           BoxSharedLink.Permissions permissions) {
 
-        BoxSharedLink sharedLink = new BoxSharedLink(access, unshareDate, permissions);
-        Info info = new Info();
-        info.setSharedLink(sharedLink);
-
-        this.updateInfo(info);
-        return info.getSharedLink();
+        return this.createSharedLink(new BoxSharedLink(access, unshareDate, permissions));
     }
 
     /**
@@ -232,12 +243,27 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @param permissions The permissions to set on the shared link for the Box folder.
      * @param password    Password set on the shared link to give access to the Box folder.
      * @return information about the newly created shared link.
+     * @deprecated Use {@link BoxFolder#createSharedLink(BoxSharedLinkRequest)}
      */
+    @Deprecated
     public BoxSharedLink createSharedLink(BoxSharedLink.Access access, Date unshareDate,
                                           BoxSharedLink.Permissions permissions, String password) {
 
-        BoxSharedLink sharedLink = new BoxSharedLink(access, unshareDate, permissions, password);
-        Info info = new Info();
+        return this.createSharedLink(new BoxSharedLink(access, unshareDate, permissions, password));
+    }
+
+    /**
+     * Creates a shared link.
+     *
+     * @param sharedLinkRequest Shared link to create
+     * @return Created shared link.
+     */
+    public BoxSharedLink createSharedLink(BoxSharedLinkRequest sharedLinkRequest) {
+        return createSharedLink(sharedLinkRequest.asSharedLink());
+    }
+
+    private BoxSharedLink createSharedLink(BoxSharedLink sharedLink) {
+        BoxFolder.Info info = new BoxFolder.Info();
         info.setSharedLink(sharedLink);
 
         this.updateInfo(info);
@@ -255,10 +281,10 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
 
         BoxAPIRequest request = new BoxAPIRequest(api, url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         int entriesCount = responseJSON.get("total_count").asInt();
-        Collection<BoxCollaboration.Info> collaborations = new ArrayList<BoxCollaboration.Info>(entriesCount);
+        Collection<BoxCollaboration.Info> collaborations = new ArrayList<>(entriesCount);
         JsonArray entries = responseJSON.get("entries").asArray();
         for (JsonValue entry : entries) {
             JsonObject entryObject = entry.asObject();
@@ -298,7 +324,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "PUT");
         request.setBody(info.getPendingChanges());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
         info.update(jsonObject);
     }
 
@@ -323,7 +349,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
 
         request.setBody(copyInfo.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
         BoxFolder copiedFolder = new BoxFolder(this.getAPI(), responseJSON.get("id").asString());
         return copiedFolder.new Info(responseJSON);
     }
@@ -346,7 +372,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             "POST");
         request.setBody(newFolder.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         BoxFolder createdFolder = new BoxFolder(this.getAPI(), responseJSON.get("id").asString());
         return createdFolder.new Info(responseJSON);
@@ -385,7 +411,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
 
         request.setBody(updateInfo.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
         BoxFolder movedFolder = new BoxFolder(this.getAPI(), responseJSON.get("id").asString());
         return movedFolder.new Info(responseJSON);
     }
@@ -541,7 +567,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         } else {
             response = (BoxJSONResponse) request.send(uploadParams.getProgressListener());
         }
-        JsonObject collection = JsonObject.readFrom(response.getJSON());
+        JsonObject collection = Json.parse(response.getJSON()).asObject();
         JsonArray entries = collection.get("entries").asArray();
         JsonObject fileInfoJSON = entries.get(0).asObject();
         String uploadedFileID = fileInfoJSON.get("id").asString();
@@ -609,7 +635,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             CREATE_WEB_LINK_URL.build(this.getAPI().getBaseURL()), "POST");
         request.setBody(newWebLink.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         BoxWebLink createdWebLink = new BoxWebLink(this.getAPI(), responseJSON.get("id").asString());
         return createdWebLink.new Info(responseJSON);
@@ -657,7 +683,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             .appendParam("direction", direction.toString());
 
         if (fields.length > 0) {
-            builder.appendParam("fields", fields).toString();
+            builder.appendParam("fields", fields);
         }
         final String query = builder.toString();
         return new Iterable<BoxItem.Info>() {
@@ -686,7 +712,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             .appendParam("direction", direction.toString());
 
         if (fields.length > 0) {
-            builder.appendParam("fields", fields).toString();
+            builder.appendParam("fields", fields);
         }
         final String query = builder.toString();
         return new Iterable<BoxItem.Info>() {
@@ -712,17 +738,17 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
             .appendParam("offset", offset);
 
         if (fields.length > 0) {
-            builder.appendParam("fields", fields).toString();
+            builder.appendParam("fields", fields);
         }
 
         URL url = GET_ITEMS_URL.buildWithQuery(getAPI().getBaseURL(), builder.toString(), getID());
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         String totalCountString = responseJSON.get("total_count").toString();
         long fullSize = Double.valueOf(totalCountString).longValue();
-        PartialCollection<BoxItem.Info> children = new PartialCollection<BoxItem.Info>(offset, limit, fullSize);
+        PartialCollection<BoxItem.Info> children = new PartialCollection<>(offset, limit, fullSize);
         JsonArray jsonArray = responseJSON.get("entries").asArray();
         for (JsonValue value : jsonArray) {
             JsonObject jsonObject = value.asObject();
@@ -832,7 +858,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         BoxJSONRequest request = new BoxJSONRequest(this.getAPI(), url, "PUT");
         request.setBody(infoJSON.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
         return new Info(jsonObject);
     }
 
@@ -872,7 +898,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         request.addHeader("Content-Type", "application/json");
         request.setBody(metadata.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -884,7 +910,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @return the metadata returned from the server.
      */
     public Metadata setMetadata(String templateName, String scope, Metadata metadata) {
-        Metadata metadataValue = null;
+        Metadata metadataValue;
 
         try {
             metadataValue = this.createMetadata(templateName, scope, metadata);
@@ -899,7 +925,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
                         metadataToUpdate.add(value.asObject().get("path").asString(),
                             value.asObject().get("value").asString());
                     } else if (value.asObject().get("value").isArray()) {
-                        ArrayList<String> list = new ArrayList<String>();
+                        ArrayList<String> list = new ArrayList<>();
                         for (JsonValue jsonValue : value.asObject().get("value").asArray()) {
                             list.add(jsonValue.asString());
                         }
@@ -946,7 +972,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         URL url = METADATA_URL_TEMPLATE.buildAlpha(this.getAPI().getBaseURL(), this.getID(), scope, templateName);
         BoxAPIRequest request = new BoxAPIRequest(this.getAPI(), url, "GET");
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -962,7 +988,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         request.addHeader("Content-Type", "application/json-patch+json");
         request.setBody(metadata.getPatch());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        return new Metadata(JsonObject.readFrom(response.getJSON()));
+        return new Metadata(Json.parse(response.getJSON()).asObject());
     }
 
     /**
@@ -1031,7 +1057,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      */
     public String setClassification(String classificationType) {
         Metadata metadata = new Metadata().add(Metadata.CLASSIFICATION_KEY, classificationType);
-        Metadata classification = null;
+        Metadata classification;
 
         try {
             classification = this.createMetadata(Metadata.CLASSIFICATION_TEMPLATE_KEY, "enterprise", metadata);
@@ -1085,7 +1111,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         request.setBody(body.toString());
 
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject jsonObject = JsonObject.readFrom(response.getJSON());
+        JsonObject jsonObject = Json.parse(response.getJSON()).asObject();
 
         String sessionId = jsonObject.get("id").asString();
         BoxFileUploadSession session = new BoxFileUploadSession(this.getAPI(), sessionId);
@@ -1196,10 +1222,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      * @return the Iterable of Box Metadata Cascade Policies in your enterprise.
      */
     public Iterable<BoxMetadataCascadePolicy.Info> getMetadataCascadePolicies(String... fields) {
-        Iterable<BoxMetadataCascadePolicy.Info> cascadePoliciesInfo =
-            BoxMetadataCascadePolicy.getAll(this.getAPI(), this.getID(), fields);
-
-        return cascadePoliciesInfo;
+        return BoxMetadataCascadePolicy.getAll(this.getAPI(), this.getID(), fields);
     }
 
     /**
@@ -1212,10 +1235,8 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
      */
     public Iterable<BoxMetadataCascadePolicy.Info> getMetadataCascadePolicies(String enterpriseID,
                                                                               int limit, String... fields) {
-        Iterable<BoxMetadataCascadePolicy.Info> cascadePoliciesInfo =
-            BoxMetadataCascadePolicy.getAll(this.getAPI(), this.getID(), enterpriseID, limit, fields);
 
-        return cascadePoliciesInfo;
+        return BoxMetadataCascadePolicy.getAll(this.getAPI(), this.getID(), enterpriseID, limit, fields);
     }
 
     /**
@@ -1242,7 +1263,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
                 "POST");
         request.setBody(body.toString());
         BoxJSONResponse response = (BoxJSONResponse) request.send();
-        JsonObject responseJSON = JsonObject.readFrom(response.getJSON());
+        JsonObject responseJSON = Json.parse(response.getJSON()).asObject();
 
         BoxFolderLock createdFolderLock = new BoxFolderLock(this.getAPI(), responseJSON.get("id").asString());
         return createdFolderLock.new Info(responseJSON);
@@ -1664,7 +1685,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         }
 
         private List<String> parseSharedLinkAccessLevels(JsonArray jsonArray) {
-            List<String> accessLevels = new ArrayList<String>(jsonArray.size());
+            List<String> accessLevels = new ArrayList<>(jsonArray.size());
             for (JsonValue value : jsonArray) {
                 accessLevels.add(value.asString());
             }
@@ -1673,7 +1694,7 @@ public class BoxFolder extends BoxItem implements Iterable<BoxItem.Info> {
         }
 
         private List<String> parseAllowedInviteeRoles(JsonArray jsonArray) {
-            List<String> roles = new ArrayList<String>(jsonArray.size());
+            List<String> roles = new ArrayList<>(jsonArray.size());
             for (JsonValue value : jsonArray) {
                 roles.add(value.asString());
             }